obj-libusb updated (to commit 930fc827)
AdbBase class is introduced. The class is  responsible for basic functionality, such as forming of packets and version management.

Feature and Features classes are auxiliary classes meant to provide feature strings in convenient format.

UsbTransport::createTransport(...) now returns std::unique_ptr<> instead of std::optional. The pointer can be empty.
Also, a variation of the function that accepts interface hint was added.